still moving up functionality, almost done
